fixed getMousePosition() calculation
Converted to floating point division to avoid small errors

+++ src/input.cpp
@@ -49,8 +49,9 @@ Vector Input::getMousePosition() {
     int width, height;
     SDL_GetWindowSize(window, &width, &height);
 
-    x /= (width / SCREEN_WIDTH);
-    y /= (height / SCREEN_HEIGHT);
+    x /= ((float)width / (float)SCREEN_WIDTH);
+    y /= ((float)height / (float)SCREEN_HEIGHT);
+
     
     return Vector(std::round(x - 0.5), std::round(y - 0.5));
 }
